account recovery functionality added and working
account recovery functionality added and working.  Response is validated and changes the background color of all input fields if incorrect data was entered or omitted.

Will need to run command on backend to install sendgrid and to install python envdot.  A .env file also needs to be created on the backend with the send grid key in it